requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --extra-index-url=https://pypi.doubanio.com/simple --index-url=https://mirrors.cloud.tencent.com/pypi/simple/ setup.py
#
--index-url https://mirrors.cloud.tencent.com/pypi/simple/
--extra-index-url https://pypi.doubanio.com/simple

alembic==1.7.7
    # via comment (setup.py)
anyio==3.6.1
    # via
    #   starlette
    #   watchgod
argon2-cffi==21.3.0
    # via passlib
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.5.2
    # via uvicorn
certifi==2022.5.18.1
    # via requests
cffi==1.15.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via uvicorn
colorlog==6.6.0
    # via comment (setup.py)
cryptography==37.0.2
    # via python-jose
dnspython==2.2.1
    # via email-validator
ecdsa==0.17.0
    # via python-jose
email-validator==1.2.1
    # via pydantic
fastapi==0.78.0
    # via comment (setup.py)
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
httptools==0.4.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via mako
mysqlclient==2.1.0
    # via comment (setup.py)
passlib[argon2]==1.7.4
    # via comment (setup.py)
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic[email]==1.9.1
    # via
    #   comment (setup.py)
    #   fastapi
python-dotenv==0.20.0
    # via
    #   comment (setup.py)
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via comment (setup.py)
pyyaml==6.0
    # via uvicorn
requests==2.27.1
    # via comment (setup.py)
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   sqlalchemy-utils
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.36
    # via
    #   alembic
    #   comment (setup.py)
    #   sqlalchemy-utils
sqlalchemy-utils==0.38.2
    # via comment (setup.py)
starlette==0.19.1
    # via fastapi
typing-extensions==4.2.0
    # via
    #   pydantic
    #   starlette
urllib3==1.26.9
    # via requests
uvicorn[standard]==0.17.6
    # via comment (setup.py)
uvloop==0.16.0
    # via uvicorn
watchgod==0.8.2
    # via uvicorn
websockets==10.3
    # via uvicorn
